requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
celery==5.3.4
redis==5.0.1
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
